Revert "fix&feat: 完成地图模块化，完成地址解析"
This reverts commit e7d4bc69f2f4200d1b35082c79ec998a22a60b69.
